refactor(parameter): moving parameter to types
