+++ go/types/parameter.go
@@ -1,7 +1,7 @@
 // Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package parameters
+package types
 
 import (
 	"github.com/gogo/protobuf/proto"
